inotify: raise correct error if server is already started in a deep repository
When path is too long to be an Unix socket address, we create a socket in a
temporary directory and link from the long path to the shorter one.
But checks in server code at startup were insufficient in this case, and used
to raise an unclear "tried linking .hg/inotify.sock to a temporary socket but
.hg/inotify.sock already exists"

+++ tests/test-inotify-issue1208.t
@@ -26,4 +26,11 @@ status
 
   $ hg status
   ? hg.pid
+
+if we try to start twice the server, make sure we get a correct error
+
+  $ hg inserve -d --pid-file=hg2.pid
+  abort: inotify-server: cannot start: socket is already bound
+  abort: child process failed to start
+  [255]
   $ kill `cat hg.pid`
